+++ patientMatcher/cli/update.py
@@ -82,10 +82,12 @@ def resources(test):
     """
     files = {}
     for key, item in PHENOTYPE_TERMS.items():
-        url = item["url"]
         destination = item["resource_path"]
+        url = item["url"]
+
         r = requests.get(url, stream=True)
         total_length = int(r.headers.get("content-length"))
+
         if test:  # read file and get its size
             files[
                 key
@@ -93,6 +95,7 @@ def resources(test):
             if total_length:
                 click.echo("file {} found at the requested URL".format(key))
             continue
+
         with open(destination, "wb") as f:  # overwrite file
             for chunk in progress.bar(
                 r.iter_content(chunk_size=1024), expected_size=(total_length / 1024) + 1